Persist default values with file
To read and write the default values the readJSON and writeJSON methods
are used, i.e. the map and list variables are persisted as JSON.
Static read and write methods only work if a script is present. If not a
restart will have the same behaviour as before.
prepareDefaultValues step should be called at the beginning of the
pipeline, which is automatically done when calling
setupCommonPipelineEnvironment step.